fix: 消除Kotlin编译警告（废弃覆盖/冗余转换/不必要安全调用）
- 为废弃API覆盖添加 @Deprecated 注解：onUse、onStateReplaced、onStacksDropped、
  hasComparatorOutput、getComparatorOutput、getBackground、getCullingShape
- 移除冗余 Elvis 运算符（ModMachineRecipes.recipeType 返回非空）
- 移除不必要类型转换 as?、.toInt()、.toLong()、.toDouble()
- ContainerItemContext.withInitial → withConstant
- World.isChunkLoaded(BlockPos) → isChunkLoaded(x, z)
- SpriteAtlasTexture.BLOCK_ATLAS_TEXTURE → 显式 Identifier
- 移除不必要 !! 和安全调用（?.）
- registryEntry → Registries.ITEM.getId()
- 对废弃父类调用添加 @Suppress("DEPRECATION")

+++ core/src/client/kotlin/ic2_120/client/screen/CokeKilnScreen.kt
@@ -90,7 +90,7 @@ class CokeKilnScreen(
             val lines = if (fluidAmount > 0) {
                 val fluid = Registries.FLUID.get(handler.sync.fluidRawId)
                 val mb = fluidAmount / DROPLETS_PER_MB
-                val capMb = (8 * FluidConstants.BUCKET / DROPLETS_PER_MB).toLong()
+                val capMb = 8 * FluidConstants.BUCKET / DROPLETS_PER_MB
                 listOf(Text.literal(fluid.defaultState.blockState.block.name.string),
                     Text.literal("${"%,d".format(mb)} / ${"%,d".format(capMb)} mB"))
             } else {
